requirements.txt
flask
gunicorn
flask_wtf
wtforms
redis==3.3.11
rq==1.1.0
flask_talisman
flask-session
Flask-JSGlue
flask-cors
Flask-Limiter
Flask-Caching
Flask-Security-Too
flask_mongoengine
flask_admin
bcrypt

matplotlib
networkx
pandas
numpy
xlrd
pubchempy
pyyaml
IPython
seaborn

pymed
crossrefapi